requirements.txt
flask==3.1.0
opencv-python==4.11.0.86
pillow==11.1.0
requests==2.32.3
numpy==2.2.4
roboflow